refactor: move useStateStore to src/store/hooks
