+++ src/components/Threads/hooks/useThreadState.ts
@@ -1,7 +1,7 @@
 import { ThreadState } from 'stream-chat';
-import { useStateStore } from './useStateStore';
 import { useThreadListItemContext } from '../ThreadList';
 import { useThreadContext } from '../ThreadContext';
+import { useStateStore } from '../../../store/';
 
 /**
  * @description returns thread state, prioritizes `ThreadListItemContext` falls back to `ThreadContext` if not former is not present
