+++ src/components/Threads/index.ts
@@ -1,3 +1,2 @@
 export * from './ThreadContext';
 export * from './ThreadList';
-export * from './hooks/useStateStore';
