+++ src/index.ts
@@ -1,5 +1,6 @@
 export * from './components';
 export * from './context';
 export * from './i18n';
+export * from './store';
 export * from './types';
 export * from './utils';
